refactor(orders): extract order domain into backend/src/orders and web/src/orders modules
