+++ web/src/components/admin/pages/EditEntryPage.tsx
@@ -5,7 +5,8 @@ import Link from "next/link";
 import Image from "next/image";
 import { useSearchParams, useRouter } from "next/navigation";
 import { uploadEntryImage, deleteEntryImage } from "@/lib/imageUpload";
-import { updateEntry, createEntry, deleteEntry, moveEntry, fetchOrderDestinations, type AdminOrderDestination } from "@/lib/api";
+import { updateEntry, createEntry, deleteEntry, moveEntry } from "@/lib/api";
+import { fetchOrderDestinations, type AdminOrderDestination } from "@/orders/api";
 import { sanitizeI18nData } from "@/lib/i18n-admin";
 import { useRestaurantStore, useCategories, useLabels } from "@/stores/restaurantStore";
 import { LABEL_COLOR_STYLES, resolveLabel } from "@/lib/label-colors";
