+++ web/src/components/menu/SelectionContent.tsx
@@ -6,7 +6,8 @@ import { useParams, useRouter, useSearchParams } from "next/navigation";
 import { QRCodeSVG } from "qrcode.react";
 import { Dialog, DialogPanel } from "@headlessui/react";
 import { getLocalizedContentValue } from "@/lib/content-presentation";
-import { ApiError, createOrderIntent, submitOrder } from "@/lib/api";
+import { ApiError } from "@/lib/api";
+import { createOrderIntent, submitOrder } from "@/orders/api";
 import { useTranslations } from "@/lib/i18n";
 import type { MenuCategory, MenuEntry } from "@/lib/types";
 import { useRestaurantStore } from "@/stores/restaurantStore";
