+++ web/src/lib/api.ts
@@ -32,13 +32,8 @@ import type {
   NormalizedModulesConfig,
   ModulesConfig,
   ImageUploadResponse,
-  SubmitOrderBody,
   SubmitOrderLine,
   SubmitOrderResponse,
-  UpdateOrderStatusBody,
-  CreateOrderIntentBody,
-  CreateOrderIntentResponse,
-  OrderIntentReviewResponse,
   CreateStaffLinkBody,
   CreatedStaffLinkResponse,
   StaffLinkSummary,
@@ -113,7 +108,7 @@ export function clearStaffSession(): void {
   }
 }
 
-async function apiFetch<T>(path: string, options: FetchOptions = {}): Promise<T> {
+export async function apiFetch<T>(path: string, options: FetchOptions = {}): Promise<T> {
   const { method = 'GET', body, headers = {}, auth = false, staff = false } = options;
 
   if (staff) {
@@ -185,37 +180,6 @@ export class ApiError extends Error {
 /** Fetch the full public catalog. Cache-bust so admin edits are visible immediately. */
 export function getCatalog() {
   return apiFetch<CatalogResponse>(`/catalog?t=${Date.now()}`);
-}
-
-/**
- * Submit an order (public, rate-limited, idempotent via idempotencyKey).
- * With a tableSessionId (waiter mode, #15) it also sends the staff session
- * header so the shared /orders route authenticates the waiter.
- */
-export function submitOrder(body: SubmitOrderBody) {
-  return apiFetch<SubmitOrderResponse>('/orders', { method: 'POST', body, staff: !!body.tableSessionId });
-}
-
-/** Create a waiter-handoff order intent (public); the token goes into the QR link. */
-export function createOrderIntent(body: CreateOrderIntentBody) {
-  return apiFetch<CreateOrderIntentResponse>('/orders/intents', { method: 'POST', body });
-}
-
-/** Load an order intent for waiter review (staff, #15). Lines reflect the current menu. */
-export function fetchOrderIntent(token: string) {
-  return apiFetch<OrderIntentReviewResponse>(`/staff/order-intents/${encodeURIComponent(token)}`, { staff: true });
-}
-
-/** Consume an intent into a real order (staff). 409: expired | consumed | stale_items. Optional table session + edited lines override. */
-export function consumeOrderIntent(token: string, opts?: { tableSessionId?: string; lines?: SubmitOrderLine[] }) {
-  const body: Record<string, unknown> = {};
-  if (opts?.tableSessionId) body.tableSessionId = opts.tableSessionId;
-  if (opts?.lines) body.lines = opts.lines;
-  return apiFetch<SubmitOrderResponse>(`/staff/order-intents/${encodeURIComponent(token)}/consume`, {
-    method: 'POST',
-    body: Object.keys(body).length > 0 ? body : undefined,
-    staff: true,
-  });
 }
 
 // ── Staff / waiter mode (#15) ────────────────────────────────
@@ -667,78 +631,6 @@ export function publishCatalog() {
     method: 'POST',
     auth: true,
   });
-}
-
-// ── Kitchen board / order destinations (#18) ────────────────────────
-
-export interface AdminOrderItemDestination {
-  id: string;
-  destinationId: string | null;
-  destinationName: string;
-  printedAt: number | null;
-}
-
-export interface AdminOrderItem {
-  id: string;
-  name: string;
-  price: number;
-  quantity: number;
-  destinations: AdminOrderItemDestination[];
-}
-
-export interface AdminOrder {
-  id: string;
-  dailyNumber: number;
-  status: 'submitted' | 'ready' | 'served' | 'rejected';
-  rejectReason: string | null;
-  createdAt: number;
-  updatedAt: number;
-  tableName: string | null;
-  submittedBy: string | null;
-  items: AdminOrderItem[];
-}
-
-export function fetchAdminOrders(day?: number) {
-  const qs = day ? `?day=${day}` : '';
-  return apiFetch<{ day: number; orders: AdminOrder[] }>(`/admin/orders${qs}`, { auth: true });
-}
-
-export function updateOrderStatus(orderId: string, body: UpdateOrderStatusBody) {
-  return apiFetch<{ ok: true; status: string }>(`/admin/orders/${encodeURIComponent(orderId)}/status`, {
-    method: 'PATCH',
-    body,
-    auth: true,
-  });
-}
-
-export function setDestinationPrinted(rowId: string, printed: boolean) {
-  return apiFetch<{ ok: true; printedAt: number | null }>(`/admin/order-item-destinations/${encodeURIComponent(rowId)}/printed`, {
-    method: 'PATCH',
-    body: { printed },
-    auth: true,
-  });
-}
-
-export interface AdminOrderDestination {
-  id: string;
-  name: string;
-  sortOrder: number;
-}
-
-export function fetchOrderDestinations() {
-  return apiFetch<{ destinations: AdminOrderDestination[] }>(`/admin/order-destinations`, { auth: true });
-}
-
-export function createOrderDestination(name: string) {
-  return apiFetch<CreatedEntryResponse>(`/admin/order-destinations`, { method: 'POST', body: { name }, auth: true });
-}
-
-export function updateOrderDestination(id: string, name: string) {
-  return apiFetch(`/admin/order-destinations/${encodeURIComponent(id)}`, { method: 'PATCH', body: { name }, auth: true });
-}
-
-export function deleteOrderDestination(id: string) {
-  return apiFetch(`/admin/order-destinations/${encodeURIComponent(id)}`, { method: 'DELETE', auth: true });
 }
 
 // ── Analytics ────────────────────────────────────────────────────────
